+++ packages/gp2-components/src/organisms/index.ts
@@ -17,7 +17,6 @@ export { default as OnboardingPageFooter } from './OnboardingPageFooter';
 export { default as OnboardingPageHeader } from './OnboardingPageHeader';
 export { default as OpenQuestionsModal } from './OpenQuestionsModal';
 export { default as OutputCard } from './OutputCard';
-export { default as OutputDetailPageHeader } from './OutputDetailPageHeader';
 export { default as PageBanner } from './PageBanner';
 export { default as ProjectCard } from './ProjectCard';
 export { default as ProjectDetailHeader } from './ProjectDetailHeader';
